refactor abstract ID definition into several files
